Refactor Streamlit app into modular package
