+++ hdf5plot/streamlit_app.py
@@ -1,0 +1,3 @@
+from .ui import main
+
+__all__ = ["main"]
